+++ spec/rails60/app/controllers/application_controller.rb
@@ -1,0 +1,5 @@
+class ApplicationController < ActionController::Base
+  def hello
+    render plain: "Hi!"
+  end
+end