+++ CHANGELOG.md
@@ -5,11 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Added a rudimentary test-suite that builds a small Rails setup, starts a container, and sends a HTTP request to it. [#36](https://github.com/kaspergrubbe/grubruby-jemalloc/pull/36)
+- Print image-sizes to our release notes and when we test newly built images [d46a59a](https://github.com/kaspergrubbe/grubruby-jemalloc/pull/36/commits/d46a59aedfadc2da9606cf57377cf2dd24a642e2)
+
 ### Changed
 - Upgraded Bundler to version `2.1.4` (was `2.1.3`), includes some minor fixes: https://github.com/bundler/bundler/releases/tag/v2.1.4 [#35](https://github.com/kaspergrubbe/grubruby-jemalloc/pull/35)
 
 ### Removed
-- Get rid of Ruby `2.3` as they are no longer supported.
+- Got rid of Ruby `2.3.*` as those versions are no longer supported.
+- Got rid of Ruby `2.5.4` as that release was a bad release causing forks on threaded-programs to fail: https://www.ruby-lang.org/en/news/2019/03/15/ruby-2-5-5-released/
 
 ## [8020] - 2020-01-03